NPCClient application added to provide system with bots for testing #26 Clean/rebuild bug fixed

+++ SlowpokeHubs/InputEvent.cs
@@ -23,7 +23,7 @@ namespace SlowpokeHubs
 
     public class InputEventPoint
     {
-        public float X { get; set; }
-        public float Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
     }
 }
